==> 20-KAGGLE-KGP310/11_titanic/13_preprocessing.ipynb <==
#### Import Libraries

import os
import re
import numpy as np
import pandas as pd

from sklearn.preprocessing import MinMaxScaler, LabelEncoder

#### Setting

SEED = 1234

#### Fetch Data

DATA_PATH = os.path.join(os.getcwd(), 'data')

##### Gender Submission

gs_df = pd.read_csv(os.path.join(DATA_PATH, 'gender_submission.csv'))
gs_df.info()

gs_df.head()

##### Train And Test Data

data_df = pd.read_csv(os.path.join(DATA_PATH, 'train.csv'))
data_df

label_df = data_df.pop('Survived')

data_df.columns = [
    'id', '티켓등급', '이름', '성별', '나이', '형제자매배우자수', 
    '부모자녀수', '티켓번호', '요금', '객실번호', '승선항'
]

#### Preprocessing Data

##### Null 값 체크

data_df.isna().sum()

##### `나이` 결측값 처리

data_df[data_df.나이.isna()]

# 이름에서 호칭 분류
data_df['호칭'] = data_df['이름'].str.extract(r'([A-Za-z]+\.)')

data_df.head()

data_df.isna().sum()

data_df.호칭.value_counts()

result = data_df.groupby('호칭')['나이'].mean()
result

# 나이값이 NaN인 컬럼의 나이값을 호칭컬럼의 평균값으로 채우기
data_df['새나이'] = data_df.apply(
    lambda row: result[row['호칭']] if pd.isna(row['나이']) else row['나이'], axis=1
)
data_df.head(20)

##### 나이값 등급화: 0-18, 19-55, 55-99

# 나이값 구간 및 레이블 설정
# 미성년자와 노약자가 살 확률과 일반 성인이 살 확률 확인 필요
age_bins = [0, 18, 55, 999]
age_class = [0, 1, 2]
data_df['나이등급'] = pd.cut(data_df['새나이'], bins=age_bins, labels=age_class, right=False)
data_df.head(20)

##### 승선항 NaN값 처리

data_df.승선항.value_counts()

# NaN값 2개를 'N'으로 처리
data_df.승선항.fillna('N', inplace=True)

##### 객실번호 NaN값 처리
* 객실번호는 중요한 항목이지만 NaN값이 너무 많다.

data_df[data_df.객실번호.isna()]['티켓등급'].value_counts()
# 객실번호가 없는 사람의 경우 티켓 등급이 3등급이 대부분임

data_df['티켓등급'].value_counts()

data_df[(data_df.티켓등급 == 3) & (data_df.객실번호.notna())]

data_df.객실번호.value_counts()

# 객실번호로 새로운 컬럼을 만듬
# 알파벳을 추출해 각 알파벳에 해당하는 컬럼을 만들고 값을 대입하고
# 가지고 있는 최대 숫자만큼 컬럼을 만들고 각각 할당해주고
# NaN값은 알파벳 X로 처리함

data_df['객실구분'] = data_df['객실번호'].apply(lambda x: re.findall(r'[A-Za-z]', str(x)) if pd.notna(x) else ['X'])
unique_letters = list(set([item for sublist in data_df.객실구분 if isinstance(sublist, list) for item in sublist]))
print(unique_letters)

for letter in unique_letters:
    data_df[f'객실구분{letter}'] = data_df.객실구분.apply(lambda x: 1 if letter in x else 0)

data_df[data_df.객실구분F==1]

# 객실번호 숫자부분 처리 
data_df['객실숫자번호'] = data_df['객실번호'].apply(lambda x: re.findall(r'\d+', str(x)) if pd.notna(x) else ['0'])

max_length = data_df['객실숫자번호'].apply(lambda x: len(x) if isinstance(x, list) else 0).max()

# 4개의 숫자컬럼을 만들고 그곳에 하났기 입력
for i in range(max_length):
    data_df[f'객실숫자번호{i+1}'] = data_df['객실숫자번호'].apply(lambda x: int(x[i]) if isinstance(x, list) and len(x) > i else 0)

data_df[data_df.객실숫자번호.apply(lambda x: len(x)==4)]

data_df.isna().sum()

##### One-hot Encoding

# 티켓등급을 One-hot Encoding 함
oh_ticket = pd.get_dummies(data_df['티켓등급'], prefix='티켓등급')
oh_ticket

# 같은방법으로 성별, 승선항, 호칭, 나이등급을 One-hot Encoding
oh_sex = pd.get_dummies(data_df['성별'], prefix='성별')
oh_embarked = pd.get_dummies(data_df['승선항'], prefix='승선항')
oh_title = pd.get_dummies(data_df['호칭'], prefix='호칭')
oh_age_level = pd.get_dummies(data_df['나이등급'], prefix='나이등급')

# data_df 추가
data_df = pd.concat([data_df, oh_ticket, oh_sex, oh_embarked, oh_title, oh_age_level], axis=1)
data_df.shape

# one-hot encoding으로 생성된 true, false값을 숫자로 변환
data_df = data_df.replace({True: 1, False: 0})

##### 티켓번호 처리
* 티켓번호를 '문숫자'와 '숫자'로 분류하여, 문숫자는 '티켓번호구분', '숫자'는 '티켓숫자' 컬럼에 저장

data_df['티켓번호구분'] = ''
data_df['티켓숫자'] = ''

for idx, value in data_df['티켓번호'].items():
    parts = value.split()
    alpha_numerics = [re.sub(r'[^A-Za-z0-9]', '', part) for part in parts]
    
    for an_value in alpha_numerics:
        if any(char.isalpha() for char in an_value):
            data_df.at[idx, '티켓번호구분'] = an_value
        elif an_value.isnumeric():
            data_df.at[idx, '티켓숫자'] = an_value

data_df.head()

data_df.티켓번호구분.value_counts()

# 티켓번호구분을 숫자로 변경
label_encoder = LabelEncoder()
data_df['티켓번호구분숫자'] = label_encoder.fit_transform(data_df.티켓번호구분)

##### 불필요 컬럼 제거

data_df.isna().sum()

# 위에서 One-Hot Encoding한 컬럼을 포함해 아래 컬럼 삭제
drop_columns = [
    'id', '티켓등급', '이름', '성별', '나이', '티켓번호', '객실번호', 
    '호칭', '승선항', '나이등급', '객실구분', '객실숫자번호', '티켓번호구분' 
]

data_df = data_df.drop(columns=drop_columns)

##### object컬럼 변환

object_columns = data_df.select_dtypes(include=['object'])
object_columns

data_df.티켓숫자 = data_df.티켓숫자.replace('', 0)
data_df.티켓숫자 = data_df.티켓숫자.astype(int)

##### 숫자데이터 정규화

scaler = MinMaxScaler()

scale_columns = ['요금', '새나이']
data_df[scale_columns] = scaler.fit_transform(data_df[scale_columns])

#### 전처리결과 데이터 저장

data_df.to_csv(
    os.path.join(DATA_PATH, 'preprocessing_data.csv'),
    index=False
)
label_df.to_csv(
    os.path.join(DATA_PATH, 'preprocessing_label.csv'),
    index=False
)
