==> 20-KAGGLE-KGP310/11_titanic/22_SVM.ipynb <==
#### 라이브러리 불러오기

import os
import pandas as pd

from sklearn.model_selection import train_test_split
from sklearn.svm import SVC
from sklearn.preprocessing import StandardScaler
from sklearn.metrics import accuracy_score

import matplotlib.pyplot as plt

##### 한글 설정

import matplotlib.font_manager as fm
import platform

os_name = platform.system()

# font_path는 사용할 폰트에 따라 변경하면 됨
font_path = 'c:/Windows/Fonts/malgun.ttf' if os_name=='Windows' \
    else '/usr/share/fonts/truetype/nanum/NanumGothic.ttf'
font_family = fm.FontProperties(fname=font_path).get_name()

# font 설정
plt.rcParams['font.family'] = font_family

# 폰트캐시까지 삭제 후 한글폰트가 사용될 수 있도록 캐시 삭제
# --> 이렇게 해야 비로소 한글이 표현되는 경우가 많음
plt.rcParams['font.family'] = 'sans-serif'
plt.rcParams['font.sans-serif'] = [font_family]

#### 원본데이터를 이용한 모델링

##### 원본데이터 불러오기

DATA_PATH = os.path.join(os.getcwd(), 'data')
data_df = pd.read_csv(os.path.join(DATA_PATH, 'train.csv'))

data_df.isna().sum()

##### 데이터 (간단) 전처리

* 결측치 처리

# 나이는 전체 중앙값으로 변환
data_df['Age'].fillna(data_df['Age'].median(), inplace=True)

# 승선항은 최빈값(mode()중 첫번째)으로 변환
data_df['Embarked'].fillna(data_df['Embarked'].mode()[0], inplace=True)

* 문자열 컬럼 숫자로 변환

data_df.Sex = data_df.Sex.map({'male':0, 'female': 1})
data_df.Embarked = data_df.Embarked.map({
    'S': 0, 'C': 1, 'Q': 2
})

##### 훈련/학습용 데이터 분리

SEED = 123

# 필요한 특성과 타겟 변수 선택
features = ['Pclass', 'Sex', 'Age', 'SibSp', 'Parch', 'Fare', 'Embarked']
target = 'Survived'

data_x = data_df[features]
data_y = data_df[target]

train_x, test_x, train_y, test_y = \
    train_test_split(data_x, data_y, test_size=0.2, random_state=SEED)

scaler = StandardScaler()
train_x = scaler.fit_transform(train_x)
test_x = scaler.fit_transform(test_x)

##### 모델 훈련

model = SVC(kernel='linear')
model.fit(train_x, train_y)

##### 정확도 확인

predictions = model.predict(test_x)
accuracy = accuracy_score(test_y, predictions)
print(f'RandomForestClassifier Accuracy: {accuracy:.5f}')

#### 전처리된 데이터를 이용한 모델링

##### 전처리 데이터 불러오기

data_df = pd.read_csv(os.path.join(DATA_PATH, 'preprocessing_data.csv'))
label_df = pd.read_csv(os.path.join(DATA_PATH, 'preprocessing_label.csv'))

##### 훈련/시험용 데이터 분리

train_x, test_x, train_y, test_y = \
    train_test_split(data_df, label_df, test_size=0.2, random_state=SEED)

scaler = StandardScaler()
train_x = scaler.fit_transform(train_x)
test_x = scaler.fit_transform(test_x)

##### 모델 훈련

model = SVC(kernel='linear')
model.fit(train_x, train_y)

##### 정확도 확인

predictions = model.predict(test_x)
accuracy = accuracy_score(test_y, predictions)
print(f'RandomForestClassifier Accuracy: {accuracy:.5f}')
